OTA: Fix a widespread NPE during state saving
The package information for the latest updates could be left unset
(null) in case the update process fails which results in a NPE being
thrown when it the value is interacted with later on. This commit
implements a simple fix to ensure that in the worst case scenario,
when the "last updates" value is requested to be set as null, it is
set as an empty array which allows it to be freely used without the
risk of a NPE being thrown.

Fixes AOSPA-642.

Change-Id: Ifb06583562e3a4891170d17622118fe1f6877dc5